refactor: converter 클래스를 util/converter로 이동
