+++ src/main/java/com/iglooclub/nungil/domain/Member.java
@@ -3,8 +3,7 @@ package com.iglooclub.nungil.domain;
 import com.iglooclub.nungil.domain.enums.*;
 import com.iglooclub.nungil.dto.ProfileCreateRequest;
 import com.iglooclub.nungil.dto.ProfileUpdateRequest;
-import com.iglooclub.nungil.util.StringListConverter;
-import com.iglooclub.nungil.util.YoilListConverter;
+import com.iglooclub.nungil.util.converter.YoilListConverter;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
